+++ ui/musicark_ui/lib/app_info.dart
@@ -2,8 +2,8 @@ class AppInfo {
   const AppInfo._();
 
   static const name = 'MusicArk';
-  static const version = '0.9.4';
-  static const backendVersion = '0.9.4';
+  static const version = '0.9.5';
+  static const backendVersion = '0.9.5';
   static const databaseSchemaVersion = '1.8.4';
   static const repositoryUrl = 'https://github.com/Regstar2/music-ark';
 }